+++ eyelid-match-ops/benches/match-ops.rs
@@ -5,6 +5,7 @@
 // TODO: move the macros to a separate module and allow missing docs only in that module.
 #![allow(missing_docs)]
 
+use eyelid_match_ops::primitives::yashe::{Yashe, YasheParams};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 
 use eyelid_match_ops::{
@@ -12,7 +13,7 @@ use eyelid_match_ops::{
         self,
         test::gen::{random_iris_code, random_iris_mask},
     },
-    primitives::poly::{self, sample_gaussian, test::gen::rand_poly, Poly, FULL_RES_POLY_DEGREE},
+    primitives::poly::{self, test::gen::rand_poly, Poly, FULL_RES_POLY_DEGREE},
 };
 
 // Configure Criterion:
@@ -229,7 +230,13 @@ pub fn bench_inv(settings: &mut Criterion) {
 
     let rng = rand::thread_rng();
 
-    let p = sample_gaussian::<FULL_RES_POLY_DEGREE>(rng);
+    let params = YasheParams {
+        t: 1024,
+        delta: 3.2,
+    };
+    let ctx: Yashe<2048> = Yashe::new(params);
+
+    let p = ctx.sample_gaussian(rng);
 
     settings.bench_with_input(
         BenchmarkId::new(
