+++ eyelid-match-ops/src/primitives/poly/test/inv.rs
@@ -1,9 +1,9 @@
 //! Tests for polynomial inverse.
 
-use crate::primitives::poly::sample_gaussian;
 use ark_ff::{One, Zero};
 
-use crate::primitives::poly::test::gen::rand_poly;
+use crate::primitives::yashe::YasheParams;
+use crate::primitives::{poly::test::gen::rand_poly, yashe::Yashe};
 use crate::primitives::poly::Poly;
 
 #[cfg(test)]
@@ -46,7 +46,13 @@ fn inverse_test_helper<const MAX_POLY_DEGREE: usize>(f: &Poly<MAX_POLY_DEGREE>) 
 #[test]
 fn test_inverse_with_small_random_coefficients() {
     let rng = rand::thread_rng();
-    let f = sample_gaussian::<FULL_RES_POLY_DEGREE>(rng);
+
+    let params = YasheParams {
+        t: 1024,
+        delta: 3.2,
+    };
+    let ctx: Yashe<2048> = Yashe::new(params);
+    let f = ctx.sample_gaussian(rng);
 
     // REMARK: For our parameter choices it is very likely to find
     // the inverse in the first attempt.
